Replaced BlockBreakingEnchant with poorly named RecursiveEnchant
