+++ src/DaPigGuy/PiggyCustomEnchants/enchants/tools/ExplosiveEnchant.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace DaPigGuy\PiggyCustomEnchants\enchants\tools;
 
 use DaPigGuy\PiggyCustomEnchants\enchants\CustomEnchant;
+use DaPigGuy\PiggyCustomEnchants\enchants\miscellaneous\RecursiveEnchant;
 use DaPigGuy\PiggyCustomEnchants\utils\PiggyExplosion;
 use pocketmine\event\block\BlockBreakEvent;
 use pocketmine\event\Event;
@@ -12,19 +13,27 @@ use pocketmine\inventory\Inventory;
 use pocketmine\item\Item;
 use pocketmine\Player;
 
-class ExplosiveEnchant extends BlockBreakingEnchant
+class ExplosiveEnchant extends RecursiveEnchant
 {
     /** @var string */
     public $name = "Explosive";
     /** @var int */
     public $rarity = CustomEnchant::RARITY_UNCOMMON;
 
+    /** @var int */
+    public $itemType = CustomEnchant::ITEM_TYPE_TOOLS;
+
+    public function getReagent(): array
+    {
+        return [BlockBreakEvent::class];
+    }
+
     public function getDefaultExtraData(): array
     {
         return ["sizeMultiplier" => 5];
     }
 
-    public function breakBlocks(Player $player, Item $item, Inventory $inventory, int $slot, Event $event, int $level, int $stack): void
+    public function safeReact(Player $player, Item $item, Inventory $inventory, int $slot, Event $event, int $level, int $stack): void
     {
         if ($event instanceof BlockBreakEvent) {
             $explosion = new PiggyExplosion($event->getBlock(), $level * $this->extraData["sizeMultiplier"], $player);
