+++ src/DaPigGuy/PiggyCustomEnchants/enchants/tools/axes/LumberjackEnchant.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace DaPigGuy\PiggyCustomEnchants\enchants\tools\axes;
 
 use DaPigGuy\PiggyCustomEnchants\enchants\CustomEnchant;
-use DaPigGuy\PiggyCustomEnchants\enchants\tools\BlockBreakingEnchant;
+use DaPigGuy\PiggyCustomEnchants\enchants\miscellaneous\RecursiveEnchant;
 use pocketmine\block\Block;
 use pocketmine\event\block\BlockBreakEvent;
 use pocketmine\event\Event;
@@ -13,7 +13,7 @@ use pocketmine\inventory\Inventory;
 use pocketmine\item\Item;
 use pocketmine\Player;
 
-class LumberjackEnchant extends BlockBreakingEnchant
+class LumberjackEnchant extends RecursiveEnchant
 {
     /** @var string */
     public $name = "Lumberjack";
@@ -23,12 +23,17 @@ class LumberjackEnchant extends BlockBreakingEnchant
     /** @var int */
     public $itemType = CustomEnchant::ITEM_TYPE_AXE;
 
+    public function getReagent(): array
+    {
+        return [BlockBreakEvent::class];
+    }
+
     public function getDefaultExtraData(): array
     {
         return ["limit" => 800];
     }
 
-    public function breakBlocks(Player $player, Item $item, Inventory $inventory, int $slot, Event $event, int $level, int $stack): void
+    public function safeReact(Player $player, Item $item, Inventory $inventory, int $slot, Event $event, int $level, int $stack): void
     {
         if ($event instanceof BlockBreakEvent) {
             $block = $event->getBlock();
@@ -37,7 +42,6 @@ class LumberjackEnchant extends BlockBreakingEnchant
                     $this->breakTree($block, $player);
                 }
             }
-            $event->setInstaBreak(true);
         }
     }
 
@@ -52,7 +56,6 @@ class LumberjackEnchant extends BlockBreakingEnchant
             if ($side->getId() !== Block::WOOD && $side->getId() !== Block::WOOD2) {
                 continue;
             }
-            $this->setCooldown($player, 1);
             $player->getLevel()->useBreakOn($side, $item, $player);
             $mined++;
             $this->breakTree($side, $player, $mined);
