+++ src/DaPigGuy/PiggyCustomEnchants/enchants/tools/hoe/HarvestEnchant.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace DaPigGuy\PiggyCustomEnchants\enchants\tools\hoe;
 
 use DaPigGuy\PiggyCustomEnchants\enchants\CustomEnchant;
-use DaPigGuy\PiggyCustomEnchants\enchants\ReactiveEnchantment;
+use DaPigGuy\PiggyCustomEnchants\enchants\miscellaneous\RecursiveEnchant;
 use pocketmine\block\Crops;
 use pocketmine\event\block\BlockBreakEvent;
 use pocketmine\event\Event;
@@ -13,7 +13,7 @@ use pocketmine\inventory\Inventory;
 use pocketmine\item\Item;
 use pocketmine\Player;
 
-class HarvestEnchant extends ReactiveEnchantment
+class HarvestEnchant extends RecursiveEnchant
 {
     /** @var string */
     public $name = "Harvest";
@@ -35,7 +35,7 @@ class HarvestEnchant extends ReactiveEnchantment
         return ["radiusMultiplier" => 1];
     }
 
-    public function react(Player $player, Item $item, Inventory $inventory, int $slot, Event $event, int $level, int $stack): void
+    public function safeReact(Player $player, Item $item, Inventory $inventory, int $slot, Event $event, int $level, int $stack): void
     {
         if ($event instanceof BlockBreakEvent) {
             $block = $event->getBlock();
@@ -44,7 +44,6 @@ class HarvestEnchant extends ReactiveEnchantment
                 for ($x = -$radius; $x <= $radius; $x++) {
                     for ($z = -$radius; $z <= $radius; $z++) {
                         if ($block->getLevel()->getBlock($block->add($x, 0, $z)) instanceof Crops) {
-                            $this->setCooldown($player, 1);
                             $block->getLevel()->useBreakOn($block->add($x, 0, $z), $item, $player);
                         }
                     }
